evaluate/requirements.txt
langchain>=0.2.0
langchain-google-genai>=0.1.12
ragas>=0.1.21
datasets>=2.19.2
python-dotenv>=1.0.1
google-generativeai>=0.5.0
